Fix issue with variable assignment and printing
Previously, the interpreter did not correctly set the value of a variable to what was inside the quotation marks when a variable was assigned. This has now been fixed, so the value is correctly set.

Additionally, when printing a variable that had been correctly assigned, the interpreter was still showing an error message of "Invalid variable". This has been resolved so that it now only shows an error message if the variable has not been assigned.